Convert Exponent to Slope Parameterization for Drug i in MuSyC Model

This can be used for setting priors and interpreting parameter estimates see MuSyC_si_to_hi() for details

Usage

MuSyC_hi_to_si(hi, Ci, E0, Ei)

Arguments

hi

numeric value of the exponent in the MuSyC equation for drug i

Ci

numeric value of the AC50 of drug i

E0

numeric value of the response with no treatments

Ei

numeric value of the response of infinite drug i and no other treatments

Value

si slope of drug i at it's AC50 and doses of all other drugs are zero

See also