Create Initial Parameter Values for the MuSyC Model

Initial values for the MuSyC model

Usage

MuSyC_init(
  logE0 = function() runif(n = 1, min = log(0.4), max = log(0.6)),
  logE1 = function() runif(n = 1, min = log(0.4), max = log(0.6)),
  logC1 = function() runif(n = 1, min = -0.2, max = 0.2),
  h1 = function() runif(n = 1, min = 3, max = 5),
  logE2 = function() runif(n = 1, min = log(0.4), max = log(0.6)),
  logC2 = function() runif(n = 1, min = -0.2, max = 0.2),
  h2 = function() runif(n = 1, min = 3, max = 5),
  logE3 = function() runif(n = 1, min = log(0.4), max = log(0.6)),
  logalpha = function() runif(n = 1, min = -0.2, max = 0.2),
  ...
)

Arguments

logE0

numeric or function returning array of length 1. Initial value for the logE0 parameter. Default: log(0.5)

logE1

numeric or function returning array of length 1. Initial value for the logE1 parameter. Default: log(0.5)

logC1

numeric or function returning array of length 1. Initial value for the logC1 parameter. Default: 0 = log(1)

h1

numeric or function returning array of length 1. Initial value for the h1 parameter. Default: 4 = MuSyC_si_to_hi(si=1, Ci=1, E0=1, Ei=0.0) the exponent for treatment 1 corresponding to a slope of 1

logE2

numeric or function returning array of length 1. Initial value for the logE2 parameter. Default: log(0.5)

logC2

numeric or function returning array of length 1. Initial value for the logC2 parameter. Default: 0 = log(1)

h2

numeric or function returning array of length 1. Initial value for the h2 parameter. Default: 4 = MuSyC_si_to_hi(si=1, Ci=1, E0=1, Ei=0.0) the exponent for treatment 1 corresponding to a slope of 1

logE3

numeric or function returning array of length 1. Initial value for the logE3 parameter. Default: log(0.5)

logalpha

numeric or function returning array of length 1. Initial value for the logalpha parameter. Default: 0 = log(1)

...

additional parameter initialization. Each named argument should be a function that returns a numeric or array depending on the dimension of the parameter, see rstan_default_init() to use the rstan default init.

Value

input for BayesPharma::model_MuSyC(init = ...) parameter.

Examples

if (FALSE) { # \dontrun{
  # default prior and init for MuSyC Model
  prior <- BayesPharma::MuSyC_prior()
  init <- BayesPharma::MuSyC_init()

  # Set prior and init for MuSyC model where both treatment 1 and treatment 2
  # are assumed to be full inhibitors
  BayesPharma::MuSyC_prior(logE0 = 0, E1 = -Inf, E2 = -Inf, E3 = -Inf)
  BayesPharma::MuSyC_init(E0 = 1, E1 = -Inf, E2 = -Inf, E3 = -Inf)
} # }