MuSyC Synergy Model

Combined: sample1 =

Usage

MuSyC_model(
  data,
  prior = MuSyC_prior(),
  init = MuSyC_init(),
  formula = MuSyC_formula(),
  control = list(adapt_delta = 0.99, max_treedepth = 12),
  stanvars = c(MuSyC_function_stanvar, MuSyC_genquant_stanvar),
  expose_functions = TRUE,
  ...
)

Arguments

data

data.frame of experimental data. must contain columns response and any predictors specified in the formula.

prior

brmspriors for kcat and kM parameters. Use BayesPharma::MuSyC_prior(...) to create the prior to use here. Default: BayesPharma::MuSyC_prior().

init

Initial values for the parameters, Use BayesPharma::MuSyC_init(...) to create the init to use here. Default BayesPharma::MuSyC_init().

formula

brmsformula object. To create a dose-response brmsformula, use the MuSyC_formula function.

control

a named list of parameters to control the sampler's behavior. Adding max_treedepth and giving a greater value than 10 can improve model convergence (default: list(adapt_delta = 0.99)).

stanvars

stanvars code for the model. Default defines the MuSyC function and genquant code.

expose_functions

boolean. Expose the BayesPharma functions for the model [default: TRUE].

...

additional arguments passed to brms::brm

Details

nlp_logE0 = X_logE0 * b_logE0 _______ _______ = 1 = param

nlp_logE1 = X_logE1 * b_logE0 _______ ______ =sample1