Generate MuSyC Ed Scores Using a Robust Functional Form

Generate MuSyC Ed Scores Using a Robust Functional Form

Usage

MuSyC_robust(
  logd1,
  logd2,
  logE0,
  logE1,
  logC1,
  h1,
  logE2,
  logC2,
  h2,
  logE3,
  logalpha
)

Arguments

logd1

numeric log dose for treatment 1

logd2

numeric log dose for treatment 2

logE0

numeric Full log response with no treatment.

logE1

numeric Full log response for treatment 1 and the dose of treatment 2 is zero.

logC1

numeric AC50 for treatment 1 and dose of treatment 2 is zero.

h1

numeric Slope of treatment 1 at its AC50 and dose of treatment 2 is zero in exponential form. Use MuSyC_si_to_hi to convert from slope form.

logE2

numeric Full log response for treatment 2 and the dose of treatment 1 is zero.

logC2

numeric AC50 for treatment 2 and dose of treatment 1 is zero.

h2

numeric Slope of treatment 2 at its AC50 and dose of treatment 1 is zero in exponential form. Use MuSyC_si_to_hi to convert from slope form.

logE3

numeric Full response for treatment 1 and treatment 2.

logalpha

numeric Log synergistic potency parameter with greater than 1 is synergistic, less than 1 is antagonistic.

Value

numeric for the synergistic response to treatment 1 and treatment 2 at doses logd1 and logd2 respectively.

References

Meyer, D.J., Wooten, D.J., Paudel B.B., Bauer, J., Hardeman, K.N., Westover, D., Lovly, C.M., Harris, L.A., Tyson D.R., Quaranta, V., Quantifying Drug Combination Synergy along Potency and Efficacy Axes, Cell Syst. 8, 2 (2019). https://doi.org/10.1016/j.cels.2019.01.003

Wooten, D.J., Meyer, C.T., Lubbock, A.L.R. et al. MuSyC is a consensus framework that unifies multi-drug synergy metrics for combinatorial drug discovery. Nat Commun 12, 4607 (2021). https://doi.org/10.1038/s41467-021-24789-z

See also