Add a log dose (in base 10) column to the input data.frame

Given the dose as a column in a data.frame with a given molar concentration add a new column of the log base-10 dose, `log_dose`, in the data.frame and return it.

calculate_log_dose(
  data,
  dose_col = dose_col,
  molar_concentration = molar_concentration)

Arguments

data
data.frame containing a column representing a dose in molar units

dose_col
expression for dose column in the input data.frame (default = dose)

molar_concentration
numeric units of molar concentration (default = 1)

Value

input data.frame with an additional `log_dose` column

Examples

# Consider observations at doses of `1 μM` and `0.1 μM`.
# If the doses are given in molar units e.g. `M` then,
data <- data.frame(dose = c(1e-6, 1e-7)) %>%
   BayesPharma::calculate_log_dose(dose_col = dose, molar_concentration = 1)
#> Error in data.frame(dose = c(1e-06, 1e-07)) %>% BayesPharma::calculate_log_dose(dose_col = dose,     molar_concentration = 1): could not find function "%>%"
data$log_dose == c(-6, -7)
#> Error in data$log_dose: object of type 'closure' is not subsettable

# If the doses are given with in units of nanomolar units e.g. `nM` then
data <- data.frame(dose_nM = c(1000, 100)) %>%
  BayesPharma::calculate_log_dose(dose_col = dose_nM, molar_concentration = 1e-9)
#> Error in data.frame(dose_nM = c(1000, 100)) %>% BayesPharma::calculate_log_dose(dose_col = dose_nM,     molar_concentration = 1e-09): could not find function "%>%"
data$log_dose == c(-6, -7)
#> Error in data$log_dose: object of type 'closure' is not subsettable