Helper function to parse command line arguments for the BayesPharma CLI

Pass in the command line arguments as a list and this will use the optparse package to parse them.

Available arguments are -v, –verbose Print extra output

–data=DATA path to either .csv, .tsv, .xls or .xlsx data file

–data_sheet=DATA_SHEET If the data file is an excel file, what sheet to use?

–data_range=DATA_RANGE If the data file is an excel file, what range to use?

–model=MODEL Which model to fit

–formula=FORMULA What type of formula to use? For models with a canonical formula, try to guess it.

–treatment_variable=TREATMENT_VARIABLE What variable to use as the treatment. If not specified use the default for the given formula.

–treatment_units=TREATMENT_UNITS What units units should be used for the treatment variable. If not specified use the default for the given formula.

–response_variable=RESPONSE_VARIABLE What variable to use as the response. If not specified use the default for the given formula.

–response_units=RESPONSE_UNITS What units units should be used for the response variable. If not specified use the default for the given formula.

–predictors=PREDICTORS Expression for the predictors for each of the model parameters.

–prior_terms=PRIOR_TERMS json formatted list of arguments for the prior function. For example, to constrain the EC50 to have a mean of 1 nM and the bottom to be 0 for for the sigmoid_agonist use --prior_terms='[{"ec50":"brms::prior(prior = normal(-9, 0.5), nlpar = "ec50")},{"bottom":"0"}]'

–init_terms=INIT_TERMS json formmated list of arguments for the init function.

–iter=ITER Number of sampling iterations.

–output_path=OUTPUT_PATH Path where the output results should be written

Usage

cli_parse_options(args)

Arguments

args

list of command line arguments

Value

named list of values for the available arguments