Create a Prior for the Michaelis Menten Enzyme Kinetics Model

For both the kcat and kM parameters, the default priors are gamma(4, 1). We use the gamma distribution because it is naturally lower bounded by 0. The first parameter is the shape, alpha=4, and the second is the rate, beta=1. The mean of gamma distributions is alpha/beta and the variance is alpha/beta^2.

Usage

michaelis_menten_prior(
  kcat = brms::prior_string(prior = "gamma(4, 1)", lb = 0, nlpar = "kcat"),
  kM = brms::prior_string(prior = "gamma(4, 1)", lb = 0, nlpar = "kM")
)

Arguments

kcat

prior for kcat parameter. Given a numeric value, it will be used as a constant and not estimated.

kM

prior for kM parameter. Given a numeric value, it will be used as a constant and not estimated.

See also