Stan Code for the Michaelis Menten Kinetics Model

The Michaelis Menten model is an ordinary differential equation model for the change in product as a function of the total enzyme concentration (ET), total substrate concentration (ST), the Michaelis constant (kM) and the catalytic constant (kcat). To implement the Michaelis Menten model in rstan::stan, the function michaelis_menten_ode is defined and then passed to michaelis_menten_single to integrate it using the stiff backward differentiation formula (BDF) method. To fit multiple time series in one model, the michaelis_menten_multiple can be used. Note that to handle fitting time-series with different numbers of observations, an additional series_index argument is used. Note that observations in the same time-series should be in sequential order in the supplied data.

Usage

michaelis_menten_stanvar()

References

Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation: Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7, 17018 (2017). https://doi.org/10.1038/s41598-017-17072-z

See also

Examples

if (FALSE) { # \dontrun{
brms::brm(
  data = ...,
  formula = brms::brmsformula(
    P ~ michaelis_menten_multiple(
      series_index, time, kcat, kM, ET, ST),
    kcat + kM ~ 1,
    nl = TRUE,
    loop = FALSE),
  prior = ...,
  init =  ...,
  stanvars = BayesPharma::michaelis_menten_stanvar())
} # }