Plot Density Distribution of Model Parameters

This function is useful to observe the distributions of the priors set for the brms::brmsfit model. By adding sample_prior = "only" to the sigmoid_model or a brms::brmsfit model, the model samples only from the prior distributions and can be observed using this plot function to verify that the values intended to be covered are being included.

Usage

plot_density_distribution(
  model,
  pars = NULL,
  labeller = NULL,
  title_label = "Density Distributions"
)

Arguments

model

`brms::brmsfit model.

pars

character vector of parameters or tidy parameter selection . using e.g. tidyselect to plot (Default: 'b_<parameter>_Intercept')

labeller

function (Default: strip off the 'b_' and '_Intercept' from the parameter labels).

title_label

character of the plot title.

Value

ggplot2::ggplot() object.

Examples

if (FALSE) { # \dontrun{
  plot_density_distribution(
    model = my_sigmoid_model)
} # }