Create Plot of a DRC Model

Create Plot of a DRC Model

Usage

plot_drc(
  model,
  newdata = NULL,
  predict_args = list(interval = "prediction"),
  aes_mapping = ggplot2::aes(x = .data[["log_dose"]], y = .data[["response"]]),
  title = NULL
)

Arguments

model

drc::drm model object generated with drc::drm()

newdata

data.frame where the predictions should be made. Default: use the range of treatment values in the model object.

predict_args

list of arguments passed to predict(model, ...) including interval will generate ribbon of the given type, see drc::predict.drc() for more details.

aes_mapping

[ggplot2::aes](...) mapping columns of the model data and new data to the plot. (Default: ggplot2::aes](x = log_dose, y = response))

title

character giving the plot title