Plot Posterior Density Distributions of modeled parameters

plots the bpfit() model density distribution of each parameter and the mean, lower confidence interval, and upper confidence interval.

Usage

plot_posterior_density(
  model,
  exclude_variables = c("__$", "lprior"),
  l_ci = 0.025,
  u_ci = 0.975,
  title_label = "Posterior Density Plots w/ Mean & 95% CI"
)

Arguments

model

bpfit model.

exclude_variables

list of regular expressions to match what variables to exclude.

l_ci

numeric unit of the lower confidence interval

u_ci

numeric unit of the upper confidence interval

title_label

string of the plot title.

Value

ggplot2::ggplot() object.

Examples

if (FALSE) { # \dontrun{
  plot_posterior_density(
    model = my_sigmoid_model,
    l_ci = 0.025,
    u_ci = 0.975,
    title_label = "Posterior Density Plots with Mean and 95% CI")
} # }