Posterior Predictive Check

pp_check from the brms package has different plot types to analyze the model fit by comparing the observed data with generated data from the model.

plot_pp_check(model, plot_type = "dens_overlay", n = 50, ...)

Arguments

model
brmsfit.

plot_type
string. Name of available plot types.

n
numeric value. Number of posterior draws to be used.

Value

ggplot object.

Details

Here is the documentation for pp_check: http://paul-buerkner.github.io/brms/reference/pp_check.brmsfit.html

Here is the documentation for pp_check plot types under PPC plotting functions': https://mc-stan.org/bayesplot/reference/PPC-overview.html

The plot function can be entered as a string for argument 'plot_type' in the function