Plot Prior and Posterior Marginal Distributions

Generate a plot that shows for each model variable the distribution of the marginal posterior uncertainty overlaid on the distribution of the marginal prior uncertainty.

Usage

plot_prior_posterior_densities(
  model,
  predictors_col_name = "_Intercept",
  half_max_label = "ec50",
  title_label = "Prior Posterior Density",
  ...
)

Arguments

model

bpfit object resulting from fitting a BayesPharma model

predictors_col_name

string expression for predictors column in the input data.frame (default = "_Intercept"). Predictors are the perturbations tested during the experiment (i.e. Drug, Temperature, etc.).

half_max_label

string of the label for the half maximal that fits the type of experiment that was done (i.e. ec50, ic50, ed50, id50, ld50, etc.).

title_label

string of the plot title.

...

further arguments passed to brms::brm() to sample from the prior

Value

ggplot2::ggplot() object.

Examples

if (FALSE) { # \dontrun{
  plot_prior_posterior_densities(
    model = my_sigmoid_model,
    predictors_col_name = "predictors",
    half_max_response = "ic50",
    title_label = "Prior Posterior Density Plots")
} # }