Plot Synergy Checkerboard

Plot Synergy Checkerboard

Usage

plot_synergy_checkerboard(
  data,
  treatment_1_variable = "dose1",
  treatment_2_variable = "dose2",
  response_variable = "response",
  treatment_1_label = "Treatment 1",
  treatment_2_label = "Treatment 2",
  treatment_1_units = NULL,
  treatment_2_units = NULL,
  plot_zero_dose = TRUE,
  contour_color = "gold"
)

Arguments

data

data.frame with columns the provided treatment_1_variable, treatment_2_variable, and response_variable

treatment_1_variable

character column name for treatment 1

treatment_2_variable

character column name for treatment 2

response_variable

character column name for the response

treatment_1_label

character used to make default title and axis labels

treatment_2_label

character used to make default title and axis labels

treatment_1_units

character used to make default axis labels

treatment_2_units

character used to make default axis labels

plot_zero_dose

logical plotted on the log scale, zero doses would be at -Inf, so to show them on the plot, add them as with a slight separation on the axis.

contour_color

character the color of the contour lines

Value

ggplot2::ggplot plot with light-blue to dark-blue tiles the dose response. Individual plot elements can be over-written and the plot can be saved with ggplot2::ggsave()