Initialize Parameter Values for an Agonist Sigmoid Model

Creating initial values for an agonist sigmoid model parameters that can be passed to the sigmoid_model() along with the sigmoid_agonist_formula() and sigmoid_agonist_prior().

Usage

sigmoid_agonist_init(
  ec50 = function() runif(n = 1, min = -7, max = -5),
  hill = function() runif(n = 1, min = 0.8, max = 1.2),
  top = function() runif(n = 1, min = 0.8, max = 1.2),
  bottom = function() runif(n = 1, min = -0.2, max = 0.2),
  ...
)

Arguments

ec50

numeric or numeric returning function units of the treatment. If the treatment is log_dose, the default value of -6 corresponds 1e-6 molar = 1 μM

hill

numeric or numeric returning function with units response/log_dose

top

numeric or numeric returning function with units of the response

bottom

numeric or numeric returning function with units of the response

...

additional parameter initialization. Each named argument should be a function that returns a numeric or array depending on the dimension of the parameter, see rstan_default_init() to use the rstan default init.

Value

input for [brms::brm](init = ...)

See also

Examples

if (FALSE) {
#Consider an activator that has a max response around 50%, EC50 is estimated
#to be around 1 nM, minimum response is known to be 0.
init <- BayesPharma::sigmoid_agonist_init(
  ec50 = -9,
  top = 0.5)
}