Create a Prior an Agonist Sigmoid Model

Creates a data.frame of priors for the sigmoid agonist model

If the function arguments ec50, hill, top, bottom are NULL, default normal distribution priors will be set.

If you would like to set a parameter as a constant, enter a numeric value for the function argument.

For other distribution options, reference brms::prior and Prior Choice Recommendations

Usage

sigmoid_agonist_prior(
  ec50 = brms::prior(prior = normal(-6, 2.5), nlpar = "ec50"),
  hill = brms::prior(prior = normal(1, 1), nlpar = "hill", lb = -0.01),
  top = brms::prior(prior = normal(1, 0.5), nlpar = "top"),
  bottom = brms::prior(prior = normal(0, 0.5), nlpar = "bottom"),
  ...
)

Arguments

ec50

brms::brmsprior or numeric. Prior for the ec50 parameter. Setting ec50 to a numeric value constrains it to a constant value. Default: normal(-6, 2.5) where the mean -6 corresponds to a concentration of 1e-6 or 1 μM.

hill

brms::brmsprior or numeric. Prior for the hill parameter. Default: normal(1, 1) with and a lower bound of -0.1.

top

brms::brmsprior or numeric giving the prior for the top parameter. Setting top to a numeric constrains it to a constant value.

bottom

brms::brmsprior or numeric giving the prior for the bottom parameter. Setting bottom to numeric constrains it to a constant value.

...

additional brms::brmsprior objects.

Value

brms::brmsprior

See also

Examples

if (FALSE) {
# Consider an activator that has a max response around 50%, EC50 is estimated
# to be around 1 nM, and minimum response is known to be 0.
priors <- sigmoid_agonist_prior(
  ec50 = brms::prior(prior = normal(-9, 0.5), nlpar = "ec50"),
  top = brms::prior(prior = normal(0.5, 0.2), nlpar = "top"),
  bottom = 0)
}