Convert Slope to Hill Parameters a the Sigmoid Functional Form

The hill parameter controls the slope at the ac50 (slope) but the slope also depends on the top and bottom parameters. This helper function facilitates computing the hill from the parameters.

To prove the this function is correct, we will re-arrange the equation relating the slope to the hill parameters

  hill * log(10) * (top - bottom) / 4 = slope
  hill = slope * 4 / (log10 * (top - bottom))

Usage

sigmoid_slope_to_hill(slope, top, bottom)

Arguments

slope

numeric the slope of the sigmoid() at the ac50

top

numeric the top parameter for the sigmoid() function.

bottom

numeric the bottom parameter for the sigmoid() function.

Value

numeric the hill coefficient of the sigmoid() function at.

See also