Stan Code for the tQ Model Generated Quantities

If only the substrate concentration is varied, it is not generally possible to fit both kcat and kM. However, it is possible to fit the ratio kcat/kM. Including this rstan::stan code will generate the samples for kcat/kM.

Usage

tQ_genquant()

References

Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation: Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7, 17018 (2017). https://doi.org/10.1038/s41598-017-17072-z

See also