Model for the tQ Enzyme Kinetics Model

Model for the tQ Enzyme Kinetics Model

Usage

tQ_model(
  data,
  formula = tQ_formula(),
  prior = tQ_prior(),
  init = tQ_init(),
  iter = 8000,
  control = list(adapt_delta = 0.99),
  stanvar_function = c(BayesPharma::tQ_stanvar(), BayesPharma::tQ_genquant()),
  expose_functions = TRUE,
  ...
)

Arguments

data

data.frame of experimental data. must contain columns time and P and any predictors specified in the formula.

formula

brms::brmsformula object. To create a dose-response brms::brmsformula, use the tQ_formula function.

prior

brms::brmsprior for kcat, and kM. Use tQ_formula() to create priors to use here.

init

list of lists, numeric value, or "random" for the initial values of the parameters being modeled.

iter

numeric of iterations the model runs. Increasing iter can help with model convergence.

control

a named list of parameters to control the sampler's behavior. Adding max_treedepth and giving a greater value than 10 can improve model convergence.

stanvar_function

stan code for the model

expose_functions

logical. Expose the BayesPharma functions for the model

...

additional arguments passed to brms::brm().

Value

bpfit object, which is a wrapper around a brms::brmsfit object.

References

Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation: Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7, 17018 (2017). https://doi.org/10.1038/s41598-017-17072-z

See also