BayesPharma: Tools for Bayesian Analysis of Non-Linear Pharmacology Models

The BayesPharma package builds on the Stan and brms to provide support for Bayesian regression modeling for foundational pharmacology models. For each model type, the user provides

formula: Describing how the model parameters, treatment, and optional predictors lead to the measured response using functions provided by BayesPharma for each model type.

observed data: response, treatment, and optional predictors as a data.frame

prior: Initial distributions over the model parameters

The models that BayesPharma support are

sigmoid_model: 4-parameter Hill equation

MuSyC_model: Bivariate synergy model with Bliss and Loewe interaction models as special cases

michaelis_menten_model: Michaelis Menten enzyme kinetics ordinary differential equation model

tQ_model: Generalization of enzyme progress curve kinetics ordinary differential equation

growth_sigmoid_model: Sigmoid model for growth kinetics

growth_richards_model: Generalized Richards model for growth kinetics

The BayesPharma package also provides a range of case studies as templates and examples for getting started at applying Bayesian modeling to pharmacology data analysis.

Building on rstan::stan brings the performance and stability of No-U-Turn Sampling (NUTs) Hamiltonian Monte Carlo and a whole ecosystem of tools for model assessment, and visualization (see https://mc-stan.org/).

Building on brms allows for compact formula based model specification adding complexity to the model incrementally, including handling missing data, measurement error, and other response distributions (see https://paul-buerkner.github.io/brms/).

See also

Author

Maintainer: Matthew O'Meara maom@umich.edu (ORCID)

Authors:

Madeline Martin mmarti29@uoregon.edu