Define a Formula for the Total QSSA (tQ) Enzyme Kinetics Model

set-up a tQ enzyme kinetics model formula to define a an ordinary differential equation model that with kcat, kM model parameters for use in the tQ_model.

Usage

tQ_formula(
  series_index_variable = "series_index",
  treatment_variable = "time",
  treatment_units = "seconds",
  ET_variable = "ET",
  ET_units = "mg/ml",
  ST_variable = "ST",
  ST_units = "mg/ml",
  response_variable = "P",
  response_units = "mg/ml",
  predictors = 1,
  ...
)

Arguments

series_index_variable

character variable indexing which measurements are part of a common time series

treatment_variable

character variable representing time as the treatment

treatment_units

character the units of the treatment variable

ET_variable

character variable representing the total enzyme concentration

ET_units

character variable representing the units of the enzyme concentration

ST_variable

character variable representing the total substrate concentration

ST_units

character variable representing the units of the substrate concentration

response_variable

character variable representing the product

response_units

character the units of the product

predictors

expression of predictors for kcat and kM parameters

...

additional arguments passed to brms::brmsformula()

Value

a bpformula, which is a subclass of brms::brmsformula and can be passed to tQ_model().

References

Choi, B., Rempala, G.A. & Kim, J.K. Beyond the Michaelis-Menten equation: Accurate and efficient estimation of enzyme kinetic parameters. Sci Rep 7, 17018 (2017). https://doi.org/10.1038/s41598-017-17072-z

See also