Initialize Parameter Values for the tQ Enzyme Kinetic Model

Initialize Parameter Values for the tQ Enzyme Kinetic Model

Usage

tQ_init(
  kcat = function() runif(n = 1, min = 3, max = 5),
  kM = function() runif(n = 1, min = 3, max = 5),
  ...
)

Arguments

kcat

numeric initial value for kcat parameter in the tQ model

kM

numeric initial value for the kM parameter in the tQ model

...

additional parameter initialization. Each named argument should be a function that returns a numeric or array depending on the dimension of the parameter, see rstan_default_init() to use the rstan default init.

Value

input for [tQ_model](init = ...) parameter.

See also